Created addcourse.php
The user is sent to this page after a course is added from the course browser. It performs a query at the beginning to check if user already has the course in their list, if not then a query is performed to add the course to the list of courses of the user. A button to go back to browser is displayed as well